\docType{data}
\name{exampleDataset}
\alias{exampleDataset}
\title{Example Dataset}
\description{
The example dataset includes randomly generated values for
3 covariates and a response variable
}
\details{
The variables included in the dataset are: \itemize{
\item\code{M} A matrix that includes cases covering three
variables \item\code{Y} A vector for the response variable
}
}

