\docType{class}
\name{LinearOutput-class}
\alias{LinearOutput-class}
\alias{getR2,LinearOutput-method}
\alias{getcoef,RegRun-method}
\alias{initialize,LinearOutput-method}
\title{A regression output object}
\description{
Objects of class\code{LinearOuput} are created by the
\code{RegRun} function
}
\details{
An object of the class 'LMOutput' has the following slots:
\itemize{ \item \code{coef} A matrix of coefficient
estimates generated by the regressions of all possible
variable combinations \item \code{R2} A vector of R-squared
values generated by the regressions of all possible
variable combinations }
}
\author{
Dino Hadzic \email{dino.hadzic@wustl.edu}
}

