\name{RegRun}
\alias{RegRun}
\alias{RegRun,ANY-method}
\title{Runs regressions on all possible combinations of covariates}
\usage{
RegRun(M, Y, ...)
}
\arguments{
  \item{M}{A matrix of cases for the covariates}

  \item{Y}{A numeric vector that whose length is equal to
  the number of rows as \code{M}}
}
\value{
An object of class LinearOuput \item{coef}{A matrix of
coefficient estimates generated by the regressions of all
possible variable combinations} \item{R2}{A vector of
R-squared values generated by the regressions of all
possible variable combinations}
}
\description{
Generates the coefficients and R-squared values on all
possible combinations of covariates
}
\examples{
exM <- matrix(rnorm(30), ncol=3)
 exY <- c(rnorm(10))
 RegRun(exM, exY)
}
\author{
Dino Hadzic \email{dino.hadzic@wustl.edu}
}
\seealso{
\code{\link{RegRunSig}}
}

